refactor: extract provider authentication logic into new ProviderConnectFlow component
